prov/efa: move progress engine code to efa_rdm_ep_progress.c
efa_rdm_ep.c is too big. This patch moved the progress
function efa_rdm_ep_progress() and related code
to a new file efa_rdm_ep_progress.c.

It also renamed some functions and added doxgen
function header
